+++ test-suite/testkit/tests/counter/counter.rs
@@ -21,19 +21,17 @@ use exonum::{
         ApiBackend,
     },
     blockchain::{IndexProof, ValidatorKeys},
-    runtime::{
-        rust::{
-            api::{ServiceApiBuilder, ServiceApiState},
-            CallContext, DefaultInstance, Service,
-        },
-        ExecutionError, InstanceId,
-    },
+    runtime::{ExecutionError, InstanceId},
 };
 use exonum_derive::*;
 use exonum_explorer::api::TransactionResponse;
 use exonum_merkledb::{
     access::{Access, FromAccess, RawAccessMut},
     ObjectHash, ProofEntry,
+};
+use exonum_rust_runtime::{
+    api::{ServiceApiBuilder, ServiceApiState},
+    CallContext, DefaultInstance, Service,
 };
 use futures::{Future, IntoFuture};
 use log::trace;