+++ CHANGELOG.md
@@ -78,10 +78,15 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
   To get the sources of an artifact, use query `type=artifact&name=$name&version=$version`,
   where `$name` and `$version` are replaced with appropriate values. (#1590)
 
+- Rust runtime module was moved from the `exonum` crate into the separate
+  `exonum-rust-runtime` crate. (#1641)
+
 - `update_service_status` now does not return a value. (#1659)
 
 - `BlockchainBuilder::build` now returns `BlockchainMut` instead of `Result`. (#1659)
+
 - A type for transaction position in block has been changed for `u32`. (#1668)
+
 - A type for a position of transaction in the block has been changed for `u32`. (#1668)
 
 #### exonum-cli
@@ -151,6 +156,8 @@ Indexes iterators names has been shortened to `Iter`, `Keys` and `Values`. (#162
 
 - Added `ProtobufConvert` for `Verified`, `BlockProof` and `IndexProof`. (#1643)
 
+- Slash (`/`) is now allowed to be a part of artifact/instance name. (#1681)
+
 #### exonum-merkledb
 
 - MerkleDB now performs automated state aggregation allowing to construct proofs
@@ -171,6 +178,10 @@ Indexes iterators names has been shortened to `Iter`, `Keys` and `Values`. (#162
 
 - `impl_serde_hex_for_binary_value` macro was moved from core to `merkledb`. (#1629)
 
+#### exonum-rust-runtime
+
+- Rust runtime module was moved from the `exonum` crate into the separate
+  `exonum-rust-runtime` crate. (#1641)
 - It is now possible to iterate over keys of the indexes within a group. (#1662)
 
 #### exonum-supervisor